feat: Analyser can handle multiple channels
Can pass in `channels` into constructor which runs a separate analysis on each channel and returns the results as a Float32Array[].

+++ Tone/component/analysis/DCMeter.ts
@@ -33,7 +33,7 @@ export class DCMeter extends MeterBase<DCMeterOptions> {
 	 * Get the signal value of the incoming signal
 	 */
 	getValue(): number {
-		const value = this._analyser.getValue();
+		const value = this._analyser.getValue() as Float32Array;
 		return value[0];
 	}
 }
